+++ api/Application/Communities/CommunityViewModel.cs
@@ -6,14 +6,14 @@ namespace Sociussion.Application.Communities;
 
 public class CommunityViewModel : IMapFrom<Community>
 {
-    public ulong Id { get; set; }
+    public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public string PhotoUrl { get; set; } = string.Empty;
 
-    public ulong OwnerId { get; set; }
+    public int OwnerId { get; set; }
 
-    public ulong MemberCount { get; set; }
+    public int MemberCount { get; set; }
 
     public DateTime? LastActive { get; set; }
     public DateTime CreatedAt { get; set; }
