+++ api/Application/Services/ICommunityService.cs
@@ -6,20 +6,20 @@ namespace Sociussion.Application.Services;
 
 public interface ICommunityService
 {
-    Task<Community> Get(ulong id);
+    Task<Community> Get(int id);
     IQueryable<Community> GetAll();
     IQueryable<Community> GetAll(CommunityQueryParams queryParams);
     
-    Task<Community> CreateFrom(CreateCommunityModel model, ulong createdBy);
+    Task<Community> CreateFrom(CreateCommunityModel model, int createdBy);
 
-    Task<bool> AddMember(ulong communityId, ulong userId);
-    Task<bool> RemoveMember(ulong communityId, ulong userId);
-    Task<CommunityViewModel> CreateFromAndGetVm(CreateCommunityModel createModel, ulong userId);
+    Task<bool> AddMember(int communityId, int userId);
+    Task<bool> RemoveMember(int communityId, int userId);
+    Task<CommunityViewModel> CreateFromAndGetVm(CreateCommunityModel createModel, int userId);
     
-    Task<Community> Update(ulong entityId, UpdateCommunityModel updateModel);
+    Task<Community> Update(int entityId, UpdateCommunityModel updateModel);
 
-    Task<CommunityViewModel> GetVm(ulong id);
+    Task<CommunityViewModel> GetVm(int id);
 
     IQueryable<CommunityViewModel> GetAllVm(CommunityQueryParams queryParams);
-    Task<bool> Delete(ulong id);
+    Task<bool> Delete(int id);
 }
