+++ api/Domain/Abstractions/IBaseEntity.cs
@@ -2,7 +2,7 @@
 
 public interface IBaseEntity
 {
-    ulong Id { get; set; }
+    int Id { get; set; }
     DateTime CreatedAt { get; set; }
     DateTime? UpdatedAt { get; set; }
 }
