+++ api/Domain/Entities/Discussion.cs
@@ -13,13 +13,13 @@ public class Discussion : BaseEntity
     public VoteScore VoteScore { get; set; } = new VoteScore();
     
     public ApplicationUser Author { get; set; }
-    public ulong? AuthorId { get; set; }
+    public int? AuthorId { get; set; }
 
     public Community Community { get; set; }
-    public ulong CommunityId { get; set; }
+    public int CommunityId { get; set; }
 
     public ICollection<Comment> Comments { get; set; }
-    public ulong CommentCount { get; set; } = 0;
+    public int CommentCount { get; set; } = 0;
 
     public ICollection<ApplicationUser> SavedBy { get; set; }
 }
